+++ src/consumer/consumer.js
@@ -554,13 +554,13 @@ export default class Consumer {
     logger.debug(`committing the following components: ${ids.join(', ')}`);
     const componentsIds = ids.map(componentId => BitId.parse(componentId));
     const components = await this.loadComponents(componentsIds);
-
     // Run over the components to check if there is missing depenedencies
     // If there is at least one we won't commit anything
     const componentsWithMissingDeps = components.filter((component) => {
       return (component.missingDependencies && !R.isEmpty(component.missingDependencies));
     });
     if (!R.isEmpty(componentsWithMissingDeps)) throw new MissingDependencies(componentsWithMissingDeps);
+
 
     const committedComponents = await this.scope
        .putMany({ consumerComponents: components, message, force, consumer: this, verbose });