+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix conflicts when import from bit.json more than one component with the same nested deps
+
 ## [0.10.3] - 2017-08-08
 
 - fix memory leak when exporting a big amount of components