+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.10.2-rc.2] - 2017-08-13
+- Use a forked version of madge for better ts support
+- Improve resolving packages dependencies (remove duplicates)
+
 ## [0.10.1] - 2017-08-07
-Improve resolving packages dependencies for ts files
+- Improve resolving packages dependencies for ts files
 
 ## [0.10.0] - 2017-08-07
 ### BREAKING CHANGES